requirements.txt
Flask-Login==0.3.2
Flask-Script==2.0.5
Flask-WTF==0.12
pymongo==3.2.1